refactor: WeatherModelプロトコルを定義し、天気取得関連処理をImplで定義
